+++ src/rules/index.ts
@@ -7,13 +7,13 @@
  */
 export type { RuleOptions } from "../typegen";
 
-export { angularRules, angularTemplateAccessibilityRules, angularTemplateRules } from "./angular";
-export { commonRules } from "./common";
-export { importRules } from "./import";
-export { javascriptRules } from "./javascript";
-export { preferLodashRules, preferLodashUnifiedRules } from "./lodash";
-export { reactRules } from "./react";
-export { packageJsonSortRules } from "./sort-package";
-export { tsconfigJsonSortRules } from "./sort-tsconfig";
-export { typescriptRules } from "./typescript";
-export { vueRules } from "./vue";
+export * from "./angular";
+export * from "./common";
+export * from "./import";
+export * from "./javascript";
+export * from "./lodash";
+export * from "./react";
+export * from "./sort-package";
+export * from "./sort-tsconfig";
+export * from "./typescript";
+export * from "./vue";